Refactoring : search in specific controller #84
