+++ test/unit/bandstorm/controller/UserControllerSpec.groovy
@@ -41,51 +41,7 @@ class UserControllerSpec extends Specification {
     }
 
 
-    void "Test the searchBand action returns a band list"() {
 
-        given: "BandDaoService exist"
-        List bandList = new ArrayList<Band>()
-        bandList.push(new Band())
-
-        Map searchResult = new HashMap()
-        searchResult.bandList = bandList
-        searchResult.bandsCount = 10
-
-        controller.bandDaoService = Mock(BandDaoService) {
-            getAllBandsByKeywords(_,_,_) >> searchResult
-        }
-
-        when: "The searchBand action is executed"
-        controller.searchBand("Bob",10,0)
-
-        then: "The bandList is correct and keywords too"
-        model.bandList
-        model.keywords == "Bob"
-
-    }
-
-    void "Test the searchUser action returns a user list"() {
-
-        given: "UserService exist"
-        List userList = new ArrayList<User>()
-        userList.push(new User(firstName:"John"))
-
-        Map searchResult = new HashMap()
-        searchResult.userList = userList
-        searchResult.userCount = 10
-
-        controller.userDaoService = Mock(UserDaoService) {
-            getAllUsersByKeywords(_,_,_) >> searchResult
-        }
-
-        when: "The searchUser action is executed"
-        controller.searchUser("John",10,0)
-
-        then: "The userList is correct and keywords too"
-        model.userList
-        model.keywords == "John"
-
-    }
 
     void "Test the create action returns the correct model"() {
         when: "The create action is executed"
@@ -94,7 +50,6 @@ class UserControllerSpec extends Specification {
         then: "The model is correctly created"
         model.userInstance != null
     }
-
 
     void "Test the passwordSettings action returns the correct model"() {
